+++ app/calendar/page.tsx
@@ -1,6 +1,7 @@
 import Layout from '@/components/Layout'
 import HabitCalendar from '@/components/HabitCalendar'
 import { ViewToggle } from '@/components/ViewToggle'
+import CompletionCountBadge from '@/components/CompletionCountBadge'
 
 export default function CalendarPage() {
   return (